feat(clientv2)!: encode requests with encoding/json/v2
Encoder used a hand-written reflect walker because encoding/json v1 has no
hook for gqlgen's MarshalGQL convention. The walker had to reimplement
struct, map, slice, pointer, omitempty and omitzero handling, and drifted
from the standard library in details such as json.Number (#328), float
formatting (#319), []byte, and embedded structs.

Go 1.27 ships encoding/json/v2, whose WithMarshalers option provides that
hook. Encoder now calls jsonv2.Marshal with DefaultOptionsV1 for v1 wire
semantics, MarshalToFunc hooks for graphql.ContextMarshaler and
graphql.Marshaler, and FormatNilSliceAsNull mapped from
NilSliceAsEmptyArray. The walker, the json.Number special case, and the
unit tests of the removed helpers are deleted, the encoder moves to its
own file, and README documents the encoding rules.

MarshalJSON now forwards its context to MarshalGQLContext, and Client
passes the request context when encoding the body.

BREAKING CHANGE: byte slices are encoded as base64 strings instead of
arrays of numbers, fields of embedded structs are flattened instead of
being nested under the type name, and struct fields are emitted in
declaration order instead of alphabetically. Use []int or a type
implementing MarshalGQL to keep sending an array.
