+++ clientv2/client.go
@@ -4,17 +4,13 @@ import (
 	"bytes"
 	"compress/gzip"
 	"context"
-	"encoding"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
 	"mime/multipart"
 	"net/http"
-	"reflect"
-	"slices"
 	"strconv"
-	"strings"
 
 	"github.com/99designs/gqlgen/graphql"
 
@@ -240,7 +236,7 @@ func (c *Client) Post(ctx context.Context, operationName, query string, respData
 
 		headers = append(headers, header{key: "Content-Type", value: contentType})
 	} else {
-		requestBody, err := c.marshalJSON(r)
+		requestBody, err := c.marshalJSON(ctx, r)
 		if err != nil {
 			return fmt.Errorf("encode: %w", err)
 		}
@@ -478,438 +474,4 @@ func (c *Client) unmarshal(data []byte, res any) error {
 	}
 
 	return err
-}
-
-func MarshalJSON(_ context.Context, v any) ([]byte, error) {
-	encoder := &Encoder{}
-	return encoder.Encode(reflect.ValueOf(v))
-}
-
-// marshalJSON encodes a request body honoring the client's encoder options.
-func (c *Client) marshalJSON(v any) ([]byte, error) {
-	encoder := &Encoder{NilSliceAsEmptyArray: c.EncodeNilSliceAsEmptyArray}
-	return encoder.Encode(reflect.ValueOf(v))
-}
-
-// Encoder is a struct for encoding GraphQL requests to JSON
-type Encoder struct {
-	// NilSliceAsEmptyArray encodes nil slices as [] instead of null.
-	// See Options.EncodeNilSliceAsEmptyArray for details.
-	NilSliceAsEmptyArray bool
-}
-
-// fieldInfo holds field information of a struct
-type fieldInfo struct {
-	name      string       // field name
-	jsonName  string       // field name in JSON
-	omitempty bool         // omitempty flag
-	omitzero  bool         // omitzero flag
-	typ       reflect.Type // field type
-}
-
-// Encode encodes any value to JSON
-func (e *Encoder) Encode(v reflect.Value) ([]byte, error) {
-	if !v.IsValid() || isNil(v) {
-		if e.NilSliceAsEmptyArray && v.IsValid() && v.Kind() == reflect.Slice {
-			return []byte("[]"), nil
-		}
-
-		return []byte("null"), nil
-	}
-
-	vi := v.Interface()
-	if marshaler, ok := vi.(graphql.ContextMarshaler); ok {
-		return e.encodeGQLContextMarshaler(context.Background(), marshaler)
-	}
-
-	if marshaler, ok := vi.(graphql.Marshaler); ok {
-		return e.encodeGQLMarshaler(marshaler)
-	}
-
-	if marshaler, ok := vi.(json.Marshaler); ok {
-		return e.encodeJsonMarshaler(marshaler)
-	}
-
-	if marshaler, ok := vi.(encoding.TextMarshaler); ok {
-		return e.encodeTextMarshaler(marshaler)
-	}
-
-	t := v.Type()
-
-	if t == jsonNumberType {
-		return e.encodeJSONNumber(v)
-	}
-
-	switch t.Kind() {
-	case reflect.Pointer:
-		return e.encodePtr(v)
-	case reflect.Struct:
-		return e.encodeStruct(v)
-	case reflect.Map:
-		return e.encodeMap(v)
-	case reflect.Slice:
-		return e.encodeSlice(v)
-	case reflect.Array:
-		return e.encodeArray(v)
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		return e.encodeInt(v)
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64, reflect.Uintptr:
-		return e.encodeUint(v)
-	case reflect.String:
-		return e.encodeString(v)
-	case reflect.Bool:
-		return e.encodeBool(v)
-	case reflect.Float32, reflect.Float64:
-		return e.encodeFloat(v)
-	case reflect.Interface:
-		return e.encodeInterface(v)
-	case reflect.Invalid, reflect.Complex64, reflect.Complex128, reflect.Chan, reflect.Func, reflect.UnsafePointer:
-		panic(fmt.Sprintf("unsupported type: %s", t))
-	default:
-		panic(fmt.Sprintf("unsupported type: %s", t))
-	}
-}
-
-// encodeGQLContextMarshaler encodes a value that implements graphql.ContextMarshaler interface
-func (e *Encoder) encodeGQLContextMarshaler(ctx context.Context, v graphql.ContextMarshaler) ([]byte, error) {
-	if isNil(reflect.ValueOf(v)) {
-		return []byte("null"), nil
-	}
-
-	var buf bytes.Buffer
-
-	err := v.MarshalGQLContext(ctx, &buf)
-	if err != nil {
-		return nil, err
-	}
-
-	return buf.Bytes(), nil
-}
-
-// encodeGQLMarshaler encodes a value that implements graphql.Marshaler interface
-func (e *Encoder) encodeGQLMarshaler(v graphql.Marshaler) ([]byte, error) {
-	if isNil(reflect.ValueOf(v)) {
-		return []byte("null"), nil
-	}
-
-	var buf bytes.Buffer
-	v.MarshalGQL(&buf)
-
-	return buf.Bytes(), nil
-}
-
-// encodeJsonMarshaler encodes a value that implements json.Marshaler interface
-func (e *Encoder) encodeJsonMarshaler(v json.Marshaler) ([]byte, error) {
-	if isNil(reflect.ValueOf(v)) {
-		return []byte("null"), nil
-	}
-
-	return v.MarshalJSON()
-}
-
-// encodeTextMarshaler encodes a value that implements encoding.TextMarshaler interface
-func (e *Encoder) encodeTextMarshaler(v encoding.TextMarshaler) ([]byte, error) {
-	if isNil(reflect.ValueOf(v)) {
-		return []byte("null"), nil
-	}
-
-	return json.Marshal(v)
-}
-
-// encodeBool encodes a boolean value
-func (e *Encoder) encodeBool(v reflect.Value) ([]byte, error) {
-	boolValue, err := json.Marshal(v.Bool())
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode bool: %v", v)
-	}
-
-	return boolValue, nil
-}
-
-// encodeInt encodes an integer value
-func (e *Encoder) encodeInt(v reflect.Value) ([]byte, error) {
-	return fmt.Appendf(nil, "%d", v.Int()), nil
-}
-
-// encodeUint encodes an unsigned integer value
-func (e *Encoder) encodeUint(v reflect.Value) ([]byte, error) {
-	return fmt.Appendf(nil, "%d", v.Uint()), nil
-}
-
-// encodeFloat encodes a floating-point value
-func (e *Encoder) encodeFloat(v reflect.Value) ([]byte, error) {
-	bitSize := 64
-	if v.Kind() == reflect.Float32 {
-		bitSize = 32
-	}
-
-	return strconv.AppendFloat(nil, v.Float(), 'g', -1, bitSize), nil
-}
-
-// encodeString encodes a string value
-func (e *Encoder) encodeString(v reflect.Value) ([]byte, error) {
-	stringValue, err := json.Marshal(v.String())
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode string: %v", v)
-	}
-
-	return stringValue, nil
-}
-
-var jsonNumberType = reflect.TypeFor[json.Number]()
-
-// encodeJSONNumber encodes a json.Number value as a JSON number.
-// json.Number does not implement json.Marshaler and its kind is reflect.String,
-// so it has to be handled before the ordinary string encoding.
-// Validation and the empty-string fallback to 0 are delegated to encoding/json.
-func (e *Encoder) encodeJSONNumber(v reflect.Value) ([]byte, error) {
-	numberValue, err := json.Marshal(json.Number(v.String()))
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode json.Number: %w", err)
-	}
-
-	return numberValue, nil
-}
-
-// trimQuotes removes double quotes from the beginning and end of a string
-func (e *Encoder) trimQuotes(s string) string {
-	if len(s) > 1 && s[0] == '"' && s[len(s)-1] == '"' {
-		return s[1 : len(s)-1]
-	}
-
-	return s
-}
-
-func isSkipField(omitempty, omitzero bool, v reflect.Value) bool {
-	if !v.IsValid() {
-		return true
-	}
-
-	var skipByOmitEmpty bool
-	if omitempty {
-		skipByOmitEmpty = isEmptyValue(v)
-	}
-
-	var skipByOmitZero bool
-	if omitzero {
-		skipByOmitZero = isZeroValue(v)
-	}
-
-	return skipByOmitEmpty || skipByOmitZero
-}
-
-// https://cs.opensource.google/go/go/+/refs/tags/go1.24.2:src/encoding/json/encode.go;l=318-330
-func isEmptyValue(v reflect.Value) bool {
-	switch v.Kind() {
-	case reflect.Array, reflect.Map, reflect.Slice, reflect.String:
-		return v.Len() == 0
-	case reflect.Bool,
-		reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64,
-		reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64, reflect.Uintptr,
-		reflect.Float32, reflect.Float64,
-		reflect.Interface, reflect.Pointer:
-		return v.IsZero()
-	}
-
-	return false
-}
-
-type isZeroer interface {
-	IsZero() bool
-}
-
-var isZeroerType = reflect.TypeFor[isZeroer]()
-
-func isZeroValue(v reflect.Value) bool {
-	// https://cs.opensource.google/go/go/+/refs/tags/go1.24.2:src/encoding/json/encode.go;l=1187-1219
-	var isZero func(v reflect.Value) bool
-
-	t := v.Type()
-	// Provide a function that uses a type's IsZero method.
-	switch {
-	case t.Kind() == reflect.Interface && t.Implements(isZeroerType):
-		isZero = func(v reflect.Value) bool {
-			// Avoid panics calling IsZero on a nil interface or
-			// non-nil interface with nil pointer.
-			return v.IsNil() ||
-				(v.Elem().Kind() == reflect.Pointer && v.Elem().IsNil()) ||
-				//nolint:forcetypeassert
-				v.Interface().(isZeroer).IsZero()
-		}
-	case t.Kind() == reflect.Pointer && t.Implements(isZeroerType):
-		//nolint:forcetypeassert
-		isZero = func(v reflect.Value) bool {
-			// Avoid panics calling IsZero on nil pointer.
-			return v.IsNil() || v.Interface().(isZeroer).IsZero()
-		}
-	case t.Implements(isZeroerType):
-		//nolint:forcetypeassert
-		isZero = func(v reflect.Value) bool {
-			return v.Interface().(isZeroer).IsZero()
-		}
-	case reflect.PointerTo(t).Implements(isZeroerType):
-		isZero = func(v reflect.Value) bool {
-			if !v.CanAddr() {
-				// Temporarily box v so we can take the address.
-				v2 := reflect.New(v.Type()).Elem()
-				v2.Set(v)
-				v = v2
-			}
-			//nolint:forcetypeassert
-			return v.Addr().Interface().(isZeroer).IsZero()
-		}
-	}
-
-	// https://cs.opensource.google/go/go/+/refs/tags/go1.24.2:src/encoding/json/encode.go;l=716
-	return isZero == nil && v.IsZero() || (isZero != nil && isZero(v))
-}
-
-// encodeStruct encodes a struct value
-func (e *Encoder) encodeStruct(v reflect.Value) ([]byte, error) {
-	fields := e.prepareFields(v.Type())
-	result := make(map[string]json.RawMessage)
-
-	for _, field := range fields {
-		fieldValue := v.FieldByName(field.name)
-		if isSkipField(field.omitempty, field.omitzero, fieldValue) {
-			continue
-		}
-
-		encodedValue, err := e.Encode(fieldValue)
-		if err != nil {
-			return nil, err
-		}
-
-		result[field.jsonName] = encodedValue
-	}
-
-	return json.Marshal(result)
-}
-
-// encodeMap encodes a map value
-func (e *Encoder) encodeMap(v reflect.Value) ([]byte, error) {
-	if v.IsNil() {
-		return []byte("null"), nil
-	}
-
-	result := make(map[string]json.RawMessage)
-
-	for _, key := range v.MapKeys() {
-		encodedKey, err := e.Encode(key)
-		if err != nil {
-			return nil, err
-		}
-
-		keyStr := string(encodedKey)
-		keyStr = e.trimQuotes(keyStr)
-
-		value := v.MapIndex(key)
-
-		encodedValue, err := e.Encode(value)
-		if err != nil {
-			return nil, err
-		}
-
-		result[keyStr] = encodedValue
-	}
-
-	return json.Marshal(result)
-}
-
-// encodeSlice encodes a slice value
-func (e *Encoder) encodeSlice(v reflect.Value) ([]byte, error) {
-	if v.IsNil() {
-		if e.NilSliceAsEmptyArray {
-			return []byte("[]"), nil
-		}
-
-		return []byte("null"), nil
-	}
-
-	return e.encodeElements(v)
-}
-
-// encodeArray encodes an array value
-func (e *Encoder) encodeArray(v reflect.Value) ([]byte, error) {
-	return e.encodeElements(v)
-}
-
-// encodeElements encodes each element of an indexable value as a JSON array
-func (e *Encoder) encodeElements(v reflect.Value) ([]byte, error) {
-	result := make([]json.RawMessage, v.Len())
-	for i := range v.Len() {
-		encodedValue, err := e.Encode(v.Index(i))
-		if err != nil {
-			return nil, err
-		}
-
-		result[i] = encodedValue
-	}
-
-	return json.Marshal(result)
-}
-
-// encodePtr encodes a pointer value
-func (e *Encoder) encodePtr(v reflect.Value) ([]byte, error) {
-	if v.IsNil() {
-		return []byte("null"), nil
-	}
-
-	return e.Encode(v.Elem())
-}
-
-// encodeInterface encodes an interface value
-func (e *Encoder) encodeInterface(v reflect.Value) ([]byte, error) {
-	if v.IsNil() {
-		return []byte("null"), nil
-	}
-
-	return e.Encode(v.Elem())
-}
-
-// prepareFields collects field information from a struct type
-func (e *Encoder) prepareFields(t reflect.Type) []fieldInfo {
-	num := t.NumField()
-
-	fields := make([]fieldInfo, 0, num)
-	for i := range num {
-		f := t.Field(i)
-		if f.PkgPath != "" && !f.Anonymous {
-			continue
-		}
-
-		jsonTag := f.Tag.Get("json")
-		if jsonTag == "-" {
-			continue
-		}
-
-		fi := fieldInfo{
-			name:     f.Name,
-			jsonName: f.Name,
-			typ:      f.Type,
-		}
-
-		if jsonTag != "" {
-			parts := strings.Split(jsonTag, ",")
-
-			fi.jsonName = parts[0]
-			if len(parts) > 1 {
-				fi.omitempty = slices.Contains(parts[1:], "omitempty")
-				fi.omitzero = slices.Contains(parts[1:], "omitzero")
-			}
-		}
-
-		fields = append(fields, fi)
-	}
-
-	return fields
-}
-
-func isNil(v reflect.Value) bool {
-	switch v.Kind() {
-	case reflect.Pointer, reflect.Map, reflect.Slice, reflect.Chan, reflect.Func, reflect.Interface:
-		return v.IsNil()
-	default:
-		return false
-	}
 }
